Refactoring: extract TreeOptimizer from SyntaxAnalyzer (#12645)
